api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.24.0
slowapi>=0.1.9
Authlib>=1.3.2
itsdangerous>=2.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
numpy==1.24.3
python-multipart==0.0.6
httpx==0.28.1
google-genai>=1.13.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyPDF2==3.0.1
python-docx==1.1.0